Tidy away multiline arg factory where it's used
