+++ lib/cucumber/core/ast.rb
@@ -6,7 +6,6 @@ require 'cucumber/core/ast/empty_multiline_argument'
 require 'cucumber/core/ast/background'
 require 'cucumber/core/ast/scenario'
 require 'cucumber/core/ast/step'
-require 'cucumber/core/ast/multiline_argument'
 require 'cucumber/core/ast/doc_string'
 require 'cucumber/core/ast/data_table'
 require 'cucumber/core/ast/scenario_outline'
